refactor: Service 합침
